bot/requirements.txt
# Python 3.10+
python-telegram-bot>=20.7
vk-api>=11.9.0
openai>=1.0.0
APScheduler>=3.10.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
openpyxl>=3.1.0
